refactor: consensus thresholds (B3) + razor compliance sweep
Phase 1 — Consensus threshold consolidation (#975):
- Production defaults: quorum=0.51, approval=0.66 (aligned with consensus_engine.py)
- ConsensusEngine now accepts approval_threshold parameter
- Hardcoded 0.5 approval check replaced with configurable threshold
- 10 threshold tests including 3-node cluster Byzantine resilience

Phase 2 — Razor compliance (7 oversized files → 18 modules ≤250 lines):
- task_execution_core.py: 1,243 → 245 + 4 sub-modules (output, reasoning, fallback, tools)
- task_reasoner.py: 555 → 162 + symbolic (191) + ml (200)
- strategy_planning.py: 421 → 202 + helpers (192)
- strategy_classification.py: 393 → 190 + heuristics (250)
- introspection_core.py: 396 → 190 + handlers (244)
- introspection_domain.py: 322 → 180 + responses (163)
- introspection_demo.py: 286 → 108 + causal (204)

Closes #975. All backlog items complete.
