+++ docs/META_LEDGER.md
@@ -330,5 +330,25 @@ SHA256(content_hash + previous_hash) = 64d536ceabad8a47a0d00383ef0a8b2dee5e8b729
 **Decision**: PASS -- Safety-critical test coverage plan approved. All 8 proposed test files target real, existing modules (4 Phase 1 safety modules verified at source paths). No production code modified. No new dependencies. No ghost paths. Two non-blocking observations: OBS-1 (auth tests must use mock/synthetic JWT keys, not real secrets), OBS-2 (Phase 3 references "23 WorldModel modules" but actual count is 54; auto-discovery pattern compensates). Implementation approved.
 
 ---
+
+### Entry #17: GATE TRIBUNAL (Consensus Threshold + Razor Compliance Plan)
+
+**Timestamp**: 2026-04-05T00:00:00Z
+**Phase**: GATE
+**Author**: Judge
+**Risk Grade**: L3
+**Verdict**: PASS
+
+**Content Hash**:
+SHA256(AUDIT_REPORT.md) = 73a720a5142b5313724f71b2139d01a12915f75005b935f4698929f0020620d2
+
+**Previous Hash**: 64d536ceabad8a47a0d00383ef0a8b2dee5e8b7297222a0a6844b3cc77de53a8
+
+**Chain Hash**:
+SHA256(content_hash + previous_hash) = abb49f197ae3e93852764a7e627503ddf604d166a3dece04be34f728301b6404
+
+**Decision**: PASS -- Consensus threshold consolidation (B3) and razor compliance sweep plan approved. All 6 audit passes clear. Phase 1: approval threshold change from 0.5 to 0.66 is security-positive and aligns canonical protocol with production ConsensusEngine; all 5 existing tests in test_consensus_protocol.py verified safe (test values at extremes 0.0 and 1.0). Phase 2: 7 oversized files (3,616 lines total) split into 18 destination modules, all at or below 250-line Section 4 Razor limit; line-count math verified against actual wc -l for all 7 source files (delta: -1 line on strategy_planning.py, rounding). No new dependencies. No orphan modules. No security stubs. Implementation approved.
+
+---
 *Chain integrity: VALID*
-*Merkle chain: 16 entries*
+*Merkle chain: 17 entries*
